gadget: Notify HAL on radio activity
Notify the HAL when the radio state changes. The behavior of RX and TX
is slightly asymmetric since the board leaves the TX state as soon as
the requested packet has been sent. This means that the gadget
interface won't notify the HAL of TX activity since only one packet is
sent (possibly multiple times).

Signed-off-by: Andreas Sandberg <andreas@sandberg.uk>