rf-preset: Add support for Ei RadioLINK bulk protocols
Add presets for Ei RadioLINK bulk protocols. This consists of two
different presets:

 * Bulk protocol for data download
 * Bulk protocol for data download (v2)

Old devices are based on the classic CC1101 radio and a discrete
microcontroller. New devices (v2) use an integrated MCU and radio
(typically a CC1310). The main difference between v1 and v2 is that
the latter uses Manchester encoding. However, the encoding used by v2
can't be processed directly in hardware on a CC1101. We therefore
perform de-whitening, Manchester decoding, and checksum calculations
in software.

Signed-off-by: Andreas Sandberg <andreas@sandberg.uk>