cc1101: Add support for back-to-back TX
The CC1101 hardware supports back-to-back TX operations without going
back to the IDLE state. The driver will support this by returning
early from send operations. A small number of outstanding send
operations can be queued in the driver, these will then be processed
back-to-back. Setting the radio driver to the IDLE or RX state acts as
a barrier that waits for TX completion.

Signed-off-by: Andreas Sandberg <andreas@sandberg.uk>