radio: Don't serialize CC1101 TX packets
The explicit transition to the IDLE state in the CC1101 driver waits
for packets to be sent. This is undesirable since this introduces a
delay between packets that are sent back-to-back. Avoid explicitly
setting the IDLE state in this case. The transition will happen
automatically once the packet has been sent.

Signed-off-by: Andreas Sandberg <andreas@sandberg.uk>